requirements.txt
llama-index==0.11.*
pyalex
python-dotenv
requests
pymupdf

llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface

ragas
bert-score
datasets
torch
transformers

streamlit

numpy<2.0.0
pandas<2.2.0